+++ client/src/components/atoms/cardImages/DeckImage.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import ButtonBase from '@material-ui/core/ButtonBase';
 import Typography from '@material-ui/core/Typography';
-import {useRecoilValue,} from 'recoil';
-import battleFieldState from '../State/battleFieldState';
+import Tooltip from '@material-ui/core/Tooltip';
+import pokeUra from '../../../image/poke_ura.jpg';
 
 const useStyles = makeStyles((theme) => ({
     image: {
@@ -75,39 +75,39 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-const BattleFieldImage = () => {
+const HandCardImage = (props) => {
     const classes = useStyles();
-    const battlePokemon = useRecoilValue(battleFieldState);
+
     return (
-        <ButtonBase
-            focusRipple
-            className={classes.image}
-            focusVisibleClassName={classes.focusVisible}
-            style={{
-                width: 130,
-            }}
-        >
-            <span
-                className={classes.imageSrc}
+        <Tooltip title={'デッキ: ' + props.deck.cards.length}>
+            <ButtonBase
+                focusRipple
+                className={classes.image}
+                focusVisibleClassName={classes.focusVisible}
                 style={{
-                    backgroundImage: `url(${battlePokemon.img_url})`,
+                    width: 130,
                 }}
-            />
-            {battlePokemon.length === 0 && (
-                <span className={classes.imageBackdrop}/>
-            )}
-            <span className={classes.imageButton}>
-                <Typography
-                    component="span"
-                    variant="subtitle1"
-                    color="inherit"
-                    className={classes.imageTitle}
-                >
-                    {'BattleField'}
-                    <span className={classes.imageMarked} />
-                </Typography>
-            </span>
-        </ButtonBase>
+            >
+                <span
+                    className={classes.imageSrc}
+                    style={{
+                        backgroundImage: `url(${pokeUra})`,
+                    }}
+                />
+                <span className={classes.imageBackdrop} />
+                <span className={classes.imageButton}>
+                    <Typography
+                        component="span"
+                        variant="subtitle1"
+                        color="inherit"
+                        className={classes.imageTitle}
+                    >
+                        {'Deck'}
+                        <span className={classes.imageMarked} />
+                    </Typography>
+                </span>
+            </ButtonBase>
+        </Tooltip>
     );
 }
-export default BattleFieldImage;
+export default HandCardImage;
