+++ client/src/components/organisms/CreatePlace.jsx
@@ -6,9 +6,9 @@ import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import InboxIcon from '@material-ui/icons/Inbox';
-import AcceptButton from './common/AcceptButton';
+import AcceptButton from '../atoms/AcceptButton';
 import Axios from 'axios';
-import UserNameContext from './Context/UserNameContext';
+import UserNameContext from '../Context/UserNameContext';
 
 const useStyles = makeStyles((theme) => ({
 	root: {
