+++ client/src/components/pages/Battle.jsx
@@ -8,9 +8,9 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import ImageButton from './common/ImageButton';
-import SelectDecks from './SelectDecks';
-import UserNameContext from './Context/UserNameContext';
+import ImageButton from '../atoms/ImageButton';
+import SelectDecks from '../molecules/SelectDecks';
+import UserNameContext from '../Context/UserNameContext';
 
 const Titles = {battle: '対戦', friendBattle: 'フレンドと対戦'}
